refactor: answer service fix, add sign check

+++ app/internal/delivery/routes/answer.py
@@ -1,19 +1,28 @@
-from typing import List
+from typing import List, Union
 
 from dependency_injector.wiring import Provide, inject
-from fastapi import APIRouter, Body, Depends, status, HTTPException
+from fastapi import APIRouter, Body, Depends, Header, HTTPException, status
 from pydantic import PositiveInt
 
+from app.config import config
+from app.core.utils import check_vk_sign
 from app.internal.deps import Application
 from app.internal.schemes import (
+    AnswerInRequest,
+    AnswerInRequestBody,
     AnswerModel,
-    CreateAnswerCommand,
     DeleteAnswerCommand,
     EmptyResult,
     GetAnswerCommand,
     Success,
     UpdateAnswerBody,
-    UpdateAnswerCommand, AnswerInRequest,
+    UpdateAnswerCommand,
+)
+from app.internal.schemes.answer_status import (
+    AnswerAlreadySent,
+    CorrectAnswer,
+    IncorrectAnswer,
+    TooManyAnswerRequests,
 )
 from app.internal.service import AnswerService
 
@@ -34,18 +43,25 @@ router = APIRouter(
 )
 @inject
 async def get_all_answers(
+    secret: str,
     skip: int = 0,
     limit: int = 100,
     answers_service: AnswerService = Depends(
         Provide[Application.services.answers_service]
     ),
 ):
+    if secret != config.BACKEND_SECRET_KEY:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized"
+        )
     return await answers_service.get_all(skip=skip, limit=limit)
 
 
 @router.post(
     "",
-    response_model=AnswerModel,
+    response_model=Union[
+        AnswerAlreadySent, TooManyAnswerRequests, CorrectAnswer, IncorrectAnswer
+    ],
     status_code=status.HTTP_201_CREATED,
     description="Create answer",
     responses={
@@ -54,15 +70,24 @@ async def get_all_answers(
 )
 @inject
 async def create_answer(
-    cmd: AnswerInRequest,
+    cmd: AnswerInRequestBody,
+    vk_params: str = Header(...),
     answers_service: AnswerService = Depends(
         Provide[Application.services.answers_service]
     ),
 ):
-    if not 1 <= cmd.task_unique_number <= 10:
-        raise HTTPException(status_code=400, detail="Task number must be between 1 and 10")
+    if not check_vk_sign(vk_params):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Bad sign")
 
-    return await answers_service.create(cmd=cmd)
+    if not 1 <= cmd.task_unique_number <= 10:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Task number must be between 1 and 10",
+        )
+
+    return await answers_service.create(
+        cmd=AnswerInRequest(vk_params=vk_params, **cmd.dict())
+    )
 
 
 @router.get(
@@ -77,31 +102,41 @@ async def create_answer(
 @inject
 async def get_answer(
     id: PositiveInt,
+    secret: str,
     answers_service: AnswerService = Depends(
         Provide[Application.services.answers_service]
     ),
 ):
+    if secret != config.BACKEND_SECRET_KEY:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized"
+        )
     return await answers_service.get(cmd=GetAnswerCommand(id=id))
 
 
-@router.put(
-    "/{id}",
-    response_model=Success,
-    status_code=status.HTTP_200_OK,
-    description="Update answer by id",
-    responses={
-        **EmptyResult().build_docs(),
-    },
-)
-@inject
-async def update_answer(
-    id: PositiveInt,
-    body: UpdateAnswerBody = Body(...),
-    answers_service: AnswerService = Depends(
-        Provide[Application.services.answers_service]
-    ),
-):
-    return await answers_service.update(cmd=UpdateAnswerCommand(id=id, **body.dict()))
+# @router.put(
+#     "/{id}",
+#     response_model=Success,
+#     status_code=status.HTTP_200_OK,
+#     description="Update answer by id",
+#     responses={
+#         **EmptyResult().build_docs(),
+#     },
+# )
+# @inject
+# async def update_answer(
+#     id: PositiveInt,
+#     secret: str,
+#     body: UpdateAnswerBody = Body(...),
+#     answers_service: AnswerService = Depends(
+#         Provide[Application.services.answers_service]
+#     ),
+# ):
+#     if secret != config.BACKEND_SECRET_KEY:
+#         raise HTTPException(
+#             status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized"
+#         )
+#     return await answers_service.update(cmd=UpdateAnswerCommand(id=id, **body.dict()))
 
 
 @router.delete(
@@ -116,29 +151,13 @@ async def update_answer(
 @inject
 async def delete_answer(
     id: PositiveInt,
+    secret: str,
     answers_service: AnswerService = Depends(
         Provide[Application.services.answers_service]
     ),
 ):
+    if secret != config.BACKEND_SECRET_KEY:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized"
+        )
     return await answers_service.delete(cmd=DeleteAnswerCommand(id=id))
-
-
-@router.post(
-    "/{id}/answer",
-    response_model=AnswerModel,
-    status_code=status.HTTP_201_CREATED,
-    description="Answer answer",
-    responses={
-        **EmptyResult().build_docs(),
-    },
-)
-@inject
-async def answer_answer(
-    id: PositiveInt,
-    cmd: CreateAnswerCommand,
-    answers_service: AnswerService = Depends(
-        Provide[Application.services.answers_service]
-    ),
-):
-    # TODO: add answer logic
-    return await answers_service.create(cmd=cmd)
